pyproject.toml
[project]
name = "phising-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bs4>=0.0.2",
    "catboost>=1.2.8",
    "flask>=3.1.0",
    "numpy>=2.0.2",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-whois>=0.9.5",
    "requests>=2.32.3",
]
